pyproject.toml
[project]
name = "nycmesh-calendar-bot"
version = "0.1"
dependencies = [
    "python-dotenv==0.19.0",
    "slack-bolt==1.9.1",
    "slack-sdk==3.10.1",
    "openai", 
    "google_api_python_client", 
    "google_auth_oauthlib", 
    "google-auth-httplib2",
    "click"
]

[project.optional-dependencies]
dev = [
    "pytest == 7.4.*",
    "flake8 == 6.1.*",
    "black == 23.7.*",
    "isort == 5.12.*",
    "coverage == 7.3.*",
    "mypy == 1.5.*",
    "watchdog"
]


[build-system]
requires = [ "setuptools >= 61", "wheel", "mypy" ]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]


[tool.mypy]
check_untyped_defs = true
show_error_codes = true
pretty = true
ignore_missing_imports = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
warn_unused_configs = true


[tool.black]
line-length = 120
include = '\.pyi?$'


[tool.isort]
profile = "black"
line_length = 120
known_first_party = [ "meshdb" ]